rename expanded -> single view
