--- _docs/expanded.md
@@ -1,4 +1,0 @@
-# Expanded View
-
-ctop provides an expanded, rolling view for following container metrics
-<p align="center"><img width="80%" src="img/expanded.gif" alt="ctop"/></p>
