+++ _docs/single.md
@@ -1,0 +1,4 @@
+# Single Container View
+
+ctop provides a rolling, single container view for following metrics
+<p align="center"><img width="80%" src="img/single.gif" alt="ctop"/></p>
